BRS-335: add sort order and featured fields
``sortOrder`` is to be used to order photos in the display (lowest
first)

``isFeatured`` is a boolean value to indicate the main photo for
search results, etc. Ideally there should be one featured photo per
park, but that is not easily enforcable in Strapi, so I think we can
fall back to sortOrder if there are multiple values.